Rename rovarspracket to AS37
Rövarspråket (100 points)
STUDENT SUBMITTED: This is a SUPER-SECRET language. Rövarspråket is not very complicated: you take an ordinary word and replace the consonants with the consonant doubled and with an "o" in between. So the consonant "b" is replaced by "bob", "r" is replaced with "ror", "s" is replaced with "sos", and so on. Vowels are left intact. It's made for Swedish, but it works just as well in English.